MAILET-115 Extract replaceMailAddresses from AbstractRedirect to MailAddressesUtils
