+++ server/mailet/mailets/src/test/java/org/apache/james/transport/mailets/redirect/AbstractRedirectTest.java
@@ -86,192 +86,6 @@ public class AbstractRedirectTest {
     }
 
     @Test
-    public void replaceMailAddressesShouldReturnEmptyWhenEmptyList() throws Exception {
-        FakeMail mail = FakeMail.builder()
-                .build();
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.<MailAddress> of());
-
-        assertThat(addresses).isEmpty();
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnSameContentWhenAddressesDoesntMatchAddressMarkerDomain() throws Exception {
-        FakeMail mail = FakeMail.builder()
-                .build();
-
-        MailAddress mailAddress = new MailAddress("user", "addres.marker");
-        MailAddress mailAddress2 = new MailAddress("user2", "address.mar");
-        ImmutableList<MailAddress> list = ImmutableList.of(mailAddress, mailAddress2);
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, list);
-
-        assertThat(addresses).containsOnly(mailAddress, mailAddress2);
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnSenderWhenAddressesMatchSender() throws Exception {
-        MailAddress sender = MailAddressFixture.ANY_AT_JAMES;
-        FakeMail mail = FakeMail.builder()
-                .sender(sender)
-                .build();
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.of(SpecialAddress.SENDER));
-
-        assertThat(addresses).containsOnly(sender);
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnSenderWhenAddressesMatchFrom() throws Exception {
-        MailAddress sender = MailAddressFixture.ANY_AT_JAMES;
-        FakeMail mail = FakeMail.builder()
-                .sender(sender)
-                .build();
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.of(SpecialAddress.FROM));
-
-        assertThat(addresses).containsOnly(sender);
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnEmptyWhenAddressesMatchSenderAndSenderIsNull() throws Exception {
-        FakeMail mail = FakeMail.builder()
-                .build();
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.of(SpecialAddress.SENDER));
-
-        assertThat(addresses).isEmpty();
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnEmptyWhenAddressesMatchReplyToAndReplyToIsNull() throws Exception {
-        MimeMessage message = new MimeMessage(Session.getDefaultInstance(new Properties()));
-        FakeMail mail = FakeMail.from(message);
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.of(SpecialAddress.REPLY_TO));
-
-        assertThat(addresses).isEmpty();
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnReplyToWhenAddressesMatchReplyTo() throws Exception {
-        MimeMessage message = new MimeMessage(Session.getDefaultInstance(new Properties()));
-        message.setReplyTo(InternetAddress.parse(MailAddressFixture.ANY_AT_JAMES.toString() + ", " + MailAddressFixture.OTHER_AT_JAMES.toString()));
-        FakeMail mail = FakeMail.from(message);
-
-        MailAddress expectedReplyTo = MailAddressFixture.ANY_AT_JAMES;
-        MailAddress expectedReplyTo2 = MailAddressFixture.OTHER_AT_JAMES;
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.of(SpecialAddress.REPLY_TO));
-
-        assertThat(addresses).containsOnly(expectedReplyTo, expectedReplyTo2);
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnSenderWhenAddressesMatchReplyToAndNoReplyTo() throws Exception {
-        MailAddress sender = MailAddressFixture.ANY_AT_JAMES;
-        FakeMail mail = FakeMail.builder()
-                .sender(sender)
-                .mimeMessage(new MimeMessage(Session.getDefaultInstance(new Properties())))
-                .build();
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.of(SpecialAddress.REPLY_TO));
-
-        assertThat(addresses).containsOnly(sender);
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnSenderWhenAddressesMatchReversePath() throws Exception {
-        MailAddress sender = MailAddressFixture.ANY_AT_JAMES;
-        FakeMail mail = FakeMail.builder()
-                .sender(sender)
-                .build();
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.of(SpecialAddress.REVERSE_PATH));
-
-        assertThat(addresses).containsOnly(sender);
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnEmptyWhenAddressesMatchReversePathAndNoSender() throws Exception {
-        FakeMail mail = FakeMail.builder()
-                .build();
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.of(SpecialAddress.REVERSE_PATH));
-
-        assertThat(addresses).isEmpty();
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnRecipientsWhenAddressesMatchRecipients() throws Exception {
-        MailAddress recipient = MailAddressFixture.ANY_AT_JAMES;
-        MailAddress recipient2 = MailAddressFixture.OTHER_AT_JAMES;
-        FakeMail mail = FakeMail.builder()
-                .recipients(recipient, recipient2)
-                .build();
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.of(SpecialAddress.RECIPIENTS));
-
-        assertThat(addresses).containsOnly(recipient, recipient2);
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnRecipientsWhenAddressesMatchTo() throws Exception {
-        MailAddress recipient = MailAddressFixture.ANY_AT_JAMES;
-        MailAddress recipient2 = MailAddressFixture.OTHER_AT_JAMES;
-        FakeMail mail = FakeMail.builder()
-                .recipients(recipient, recipient2)
-                .build();
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.of(SpecialAddress.TO));
-
-        assertThat(addresses).containsOnly(recipient, recipient2);
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnEmptyWhenAddressesMatchUnaltered() throws Exception {
-        FakeMail mail = FakeMail.builder()
-                .build();
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.of(SpecialAddress.UNALTERED));
-
-        assertThat(addresses).isEmpty();
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnEmptyWhenAddressesMatchNull() throws Exception {
-        FakeMail mail = FakeMail.builder()
-                .build();
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.of(SpecialAddress.NULL));
-
-        assertThat(addresses).isEmpty();
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnSameAddressWhenAddressesDoesntMatch() throws Exception {
-        FakeMail mail = FakeMail.builder()
-                .build();
-
-        MailAddress address = new MailAddress("user", "address.marker");
-        MailAddress address2 = new MailAddress("user2", "address.marker");
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.of(address, address2));
-
-        assertThat(addresses).containsOnly(address, address2);
-    }
-
-    @Test
-    public void replaceMailAddressesShouldReturnSameListWhenAddressesMatchDelete() throws Exception {
-        FakeMail mail = FakeMail.builder()
-                .build();
-
-        Collection<MailAddress> addresses = testee.replaceMailAddresses(mail, ImmutableList.of(SpecialAddress.DELETE));
-
-        MailAddress expected = new MailAddress("delete", "address.marker");
-        assertThat(addresses).containsOnly(expected);
-    }
-
-    @Test
     public void replaceInternetAddressesShouldReturnEmptyWhenEmptyList() throws Exception {
         FakeMail mail = FakeMail.builder()
                 .build();
